Update một chút Book
