+++ frontend/TH-ebook/package.json
@@ -11,9 +11,11 @@
   },
   "dependencies": {
     "autoprefixer": "^10.4.20",
+    "axios": "^1.7.7",
     "postcss": "^8.4.47",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-query": "^3.39.3",
     "swiper": "^11.1.14",
     "tailwindcss": "^3.4.13"
   },
