+++ frontend/TH-ebook/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 // import App from './App.tsx'
 import './index.css'
-import HungApp from './HungApp.tsx'
+import HungApp from './Book.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
